+++ lib/neo4j.rb
@@ -43,6 +43,9 @@ require 'neo4j/active_rel/related_node'
 require 'neo4j/active_rel/types'
 require 'neo4j/active_rel'
 
+require 'neo4j/active_node/dependent'
+require 'neo4j/active_node/dependent/query_proxy_methods'
+require 'neo4j/active_node/dependent/association_methods'
 require 'neo4j/active_node/query_methods'
 require 'neo4j/active_node/query/query_proxy_methods'
 require 'neo4j/active_node/query/query_proxy_enumerable'
