+++ lib/neo4j/active_node.rb
@@ -39,6 +39,7 @@ module Neo4j
     include Neo4j::ActiveNode::Rels
     include Neo4j::ActiveNode::HasN
     include Neo4j::ActiveNode::Scope
+    include Neo4j::ActiveNode::Dependent
 
     def neo4j_obj
       _persisted_obj || fail('Tried to access native neo4j object on a non persisted object')
